Add initial GitHub action for CI
This replaces the original Travis-CI configuration and allows the code
to be built and tested on the GitHub CI infrastructure.

I've based the build and test steps upon the original Travis-CI
configuration to maintain consistency with the previous setup.  This is
why `Dist::Zilla` and `Pod::Weaver::Section::Contributors` are installed
explicilty.